+++ docs/ChangeLog.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename **Message::getHeader** method to **Message::header**.
 - Rename **Message::getStorage** method to **Message::buffer**.
 
+### Removed
+
+- **Storage::canonicalize** method.
+
 ## [5.0.0] - 2021-11-18
 
 ### Added